+++ com.unity.ml-agents/CHANGELOG.md
@@ -33,6 +33,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - `StackingSensor` was changed from `internal` visibility to `public`
  - Updated Barracuda to 0.6.3-preview.
 
+ ### Bug Fixes
+ - Fixed a display bug when viewing Demonstration files in the inspector. The shapes of the observations in the file now display correctly. (#3771)
+
 ## [0.15.1-preview] - 2020-03-30
 ### Bug Fixes
  - Raise the wall in CrawlerStatic scene to prevent Agent from falling off. (#3650)